27.10

+++ GameShop/Model/ModelTableInDataBase/Product.cs
@@ -1,9 +1,4 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace GameShop.Model
 {
@@ -13,9 +8,9 @@ namespace GameShop.Model
 
         public int idCategory { get; set; }
 
-        public double Price { get; set; }
+        public float Price { get; set; }
 
-        public double Quantity { get; set; }
+        public float Quantity { get; set; }
 
         public string Name { get; set; }
